New from_list method, general improvements to ValidatedModel and ChoicesEnum.
New classmethod from_list on ChoicesEnum to enumerate from supplied arguments.
Note that alterations to the list order will break the enum. This is noted in the README.

ValidatedModel makes addErr a object method, as 'self' was previously missing.
It now only adds errors to the object's internal errs dictionary, instead of adding to a supplied dictionary.

Renamed 'name' to 'self' in ChoicesEnum.choices()
ChoicesEnum.__getattr__() now raises AttributeError if the value cannot be found.
PEP8 compliance changes.